move tests to current file
